+++ OwnMyWay/OwnMyWay/Extension/Date+.swift
@@ -66,12 +66,44 @@ extension Date {
 
 }
 
-extension Date: Strideable {
-    public func distance(to other: Date) -> TimeInterval {
-        return other.timeIntervalSinceReferenceDate - self.timeIntervalSinceReferenceDate
+extension Date {
+    var previousMonth: Date {
+        Calendar.current.date(byAdding: .month, value: -1, to: self) ?? Date()
     }
 
-    public func advanced(by interval: TimeInterval) -> Date {
-        return self + interval
+    var nextMonth: Date {
+        Calendar.current.date(byAdding: .month, value: 1, to: self) ?? Date()
     }
+
+    var weekday: Int {
+        Calendar.current.component(.weekday, from: self)
+    }
+
+    var dayNumber: Int {
+        Calendar.current.component(.day, from: self)
+    }
+
+    var month: Int {
+        Calendar.current.component(.month, from: self)
+    }
+
+    var year: Int {
+        Calendar.current.component(.year, from: self)
+    }
+
+    var firstWeekDayCount: Int {
+        return 8 - self.firstDayOfTheMonth.weekday
+    }
+
+    var numberOfDays: Int {
+        let days = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31]
+        return self.month == 2 ? self.year.isMultiple(of: 4) ? 29 : 28 : days[self.month - 1]
+    }
+
+    var firstDayOfTheMonth: Date {
+        Calendar.current.date(
+            from: Calendar.current.dateComponents([.year, .month], from: self)
+        ) ?? Date()
+    }
+
 }
